fix(barnard-rn/android, spec): address codex review batch 5
1. RN Android DetectionEvent.enin: emit as Int (not Double) so it
   validates against v2 schema's type: integer. ENIN fits in Int32 for
   ~40k years. Matches Flutter-native bridge behaviour.

2. spec 004 §2.4: correct the v2 event-scoping claim. B004 is still
   served for GAEN compat and future use, but v2 does not surface the
   peer EventCodeHash on DetectionEvent and does not filter by it.
   Mark event-scoped filtering as a follow-up, not a v2 capability.

3. spec 004 §7.1: document getDebugBuffer() and getRssiSamples() on
   BarnardClient. They were omitted from the API surface listing
   despite being shipped.

+++ packages/react-native/barnard/android/src/main/kotlin/network/greeting/barnard/BarnardController.kt
@@ -572,7 +572,10 @@ internal class BarnardController(
         }
 
         val reporterPayload = computePayload(timestampMs)
-        val enin = BarnardCrypto.calculateEnin(timestampMs).toLong()
+        // ENIN fits comfortably in Int32 for the next ~40 000 years, so
+        // prefer `putInt` to match the schema's `type: integer` contract
+        // and the wire shape of the Flutter-native bridge.
+        val enin = BarnardCrypto.calculateEnin(timestampMs).toInt()
 
         val payload = Arguments.createMap().apply {
             putString("type", "detection")
@@ -582,7 +585,7 @@ internal class BarnardController(
             putString("rpid", payloadBytes.toHex())
             putString("reporterRpid", reporterPayload.toHex())
             if (detectedDisplayIdHex != null) putString("detectedDisplayId", detectedDisplayIdHex) else putNull("detectedDisplayId")
-            putDouble("enin", enin.toDouble())
+            putInt("enin", enin)
             putInt("rssi", rssi)
             putNull("rssiSummary")
             putString("payloadRaw", payloadBytes.toHex())
